Add files via upload
Could attending university in a foreign country impact one's mental well-being? In 2018, a Japanese international university conducted a survey among its students, followed by a study the next year, which received approval from various ethical and regulatory bodies.

The study revealed that international students face a heightened risk of mental health challenges compared to the general populace. It identified social connectedness (the sense of belonging to a social group) and acculturative stress (the stress stemming from adapting to a new culture) as factors predictive of depression.

To investigate further, I delved into the student data using SQL to ascertain if I would arrive at a similar conclusion regarding international students and to determine if the duration of their stay played a role in these findings.